Solve few commnets
