+++ src/components/Portfolio/TokenImage.tsx
@@ -2,21 +2,19 @@ import type { Token } from '@lifi/widget';
 import Image from 'next/image';
 
 function TokenImage({ token }: { token: Pick<Token, 'logoURI' | 'name'> }) {
+  if (!token?.logoURI) {
+    return token.name?.slice(0, 1) || '?';
+  }
+
   return (
-    <>
-      {!token?.logoURI ? (
-        token.name?.slice(0, 1) || '?'
-      ) : (
-        <Image
-          width={0}
-          height={0}
-          sizes="100vw"
-          style={{ width: '100%', height: '100%' }} // optional
-          src={token.logoURI}
-          alt={token.name}
-        />
-      )}
-    </>
+    <Image
+      width={0}
+      height={0}
+      sizes="100vw"
+      style={{ width: '100%', height: '100%' }} // optional
+      src={token.logoURI}
+      alt={token.name}
+    />
   );
 }
