+++ src/types/fp.ts
@@ -1,5 +1,3 @@
-// ----------------------------------------------------------------------
-
 export interface FpProps {
   fp: string;
 }
